+++ CHANGELOG.md
@@ -13,16 +13,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Declared `NodeMetadataPayload.operator_signature` as `recoverable::Signature` instead of just a byte array. This allows the user to detect an invalid signature on `NodeMetadata` creation. ([#11])
 - Renamed `NodeMetadataPayload.certificate_bytes` to `certificate_der` (although it is not deserialized on the Rust side, so the DER format is not strictly enforced). ([#13])
 - Changed some method and field names in WASM bindings to conform to JS style (camel case). New names are: `TreasureMap.publisherVerifyingKey`, `TreasureMap.bobVerifyingKey`, `TreasureMap.encryptedKfrag`, `RetrievalKit.queriedAddresses`, `RevocationOrder.verifySignature`, `NodeMetadataPayload.verifyingKey`, `NodeMetadataPayload.encryptingKey`, `NodeMetadataPayload.timestampEpoch`, `MetadataRequest.announceNodes`. ([#9])
+- Moved `ADDRESS_SIZE` to `Address::SIZE`. ([#14])
+- `MetadataResponse::verify()` and `ReencryptionResponse::verify()` return a `Result` instead of `Option`. ([#14])
+- Renamed `RevocationOrder::verify_signature()` to `verify()` and made it return a `Result<(Address, EncryptedKeyFrag)>`. ([#14])
 
 
 ### Added
 
 - `TreasureMap::make_revocation_orders()` (with the corresponding methods in Python and WASM bindings). ([#9])
-- `HRAC.fromBytes()` in WASM bindings. ([#9])
+- `HRAC.fromBytes()` in WASM bindings ([#9]), and in Python bindings ([#14]).
 - `RevocationOrder.stakingProviderAddress` in WASM bindings. ([#9])
 - `MetadataResponse.verify()` in WASM bindings. ([#9])
 - `impl From<[u8; 16]>` for `HRAC`. ([#9])
-- Made `RevocationOrder.staking_provider_address` public. ([#9])
+- Made `RevocationOrder.staking_provider_address` public. ([#9]) Rolled back in ([#14]) in favor of the return value from `verify()`.
+- `HRAC::SIZE` constant ([#14])
+- `VerificationError` for use in various `verify()` methods. ([#14])
 
 
 ### Fixed
@@ -34,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#10]: https://github.com/nucypher/nucypher-core/pull/10
 [#11]: https://github.com/nucypher/nucypher-core/pull/11
 [#13]: https://github.com/nucypher/nucypher-core/pull/13
+[#14]: https://github.com/nucypher/nucypher-core/pull/14
 
 
 ## [0.0.4] - 2022-02-09